Add files via upload
Added a command that allows the user to view the contents of a file in the current working directory. When the command is selected, the program should display a list of filenames and a prompt for the name of the file to be viewed. Includes error recovery in the program. If the user enters a filename that does not exist they should be prompted to enter a filename that does exist.
An example of the program input and output is shown below:
"
/root/sandbox
1   List the current directory
2   Move up
3   Move down
4   Number of files in the directory
5   Size of the directory in bytes
6   Search for a file name
7   View the contents of a file
8   Quit the program
Enter a number: 7

Files in /root/sandbox:
filesys.py
example.txt
Enter a file name from these names: example.txt

THIS IS CONTENT OF AN EXAMPLE FILE.

/root/sandbox
1   List the current directory
2   Move up
3   Move down
4   Number of files in the directory
5   Size of the directory in bytes
6   Search for a file name
7   View the contents of a file
8   Quit the program
Enter a number: 8

Have a nice day!
"